+++ module-apps/application-calllog/CalllogModel.hpp
@@ -26,7 +26,7 @@ class CalllogModel : public app::DatabaseModel<CalllogRecord>, public gui::ListI
     CalllogModel(app::Application *app);
 
     // virtual methods
-    bool updateRecords(std::unique_ptr<std::vector<CalllogRecord>> records) override;
+    bool updateRecords(std::vector<CalllogRecord> records) override;
     void requestRecords(const uint32_t offset, const uint32_t limit) override;
 
     // virtual methods for ListViewProvider