+++ module-apps/application-desktop/ApplicationDesktop.cpp
@@ -92,8 +92,9 @@ namespace app
             }
         }
 
-        windows[app::window::name::desktop_menu]->rebuild();
-        windows[app::window::name::desktop_main_window]->rebuild();
+        /// TODO if current window is this one or was on stack -> requild it
+        /// windowsFactory.build(this, app::window::name::desktop_menu);
+        /// windowsFactory.build(this, app::window::name::desktop_main_window);
         return true;
     }
 
@@ -112,7 +113,7 @@ namespace app
              msg->interface == db::Interface::Name::SMS) &&
             msg->type != db::Query::Type::Read) {
             requestNotReadNotifications();
-            windows[app::window::name::desktop_menu]->rebuild();
+            windowsFactory.build(this, app::window::name::desktop_menu);
         }
 
         return false;
@@ -209,13 +210,24 @@ namespace app
     void ApplicationDesktop::createUserInterface()
     {
         using namespace app::window::name;
-        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_main_window, new gui::DesktopMainWindow(this)));
-        windows.insert(std::pair<std::string, gui::AppWindow *>(
-            desktop_pin_lock, new gui::PinLockWindow(this, desktop_pin_lock, lockHandler.lock)));
-        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_menu, new gui::MenuWindow(this)));
-        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_poweroff, new gui::PowerOffWindow(this)));
-        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_locked, new gui::LockedInfoWindow(this)));
-        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_reboot, new gui::RebootWindow(this)));
+        windowsFactory.attach(desktop_main_window, [](Application *app, const std::string &name) {
+            return std::make_unique<gui::DesktopMainWindow>(app);
+        });
+        windowsFactory.attach(desktop_pin_lock, [&](Application *app, const std::string newname) {
+            return std::make_unique<gui::PinLockWindow>(app, desktop_pin_lock, lockHandler.lock);
+        });
+        windowsFactory.attach(desktop_menu, [](Application *app, const std::string newname) {
+            return std::make_unique<gui::MenuWindow>(app);
+        });
+        windowsFactory.attach(desktop_poweroff, [](Application *app, const std::string newname) {
+            return std::make_unique<gui::PowerOffWindow>(app);
+        });
+        windowsFactory.attach(desktop_locked, [](Application *app, const std::string newname) {
+            return std::make_unique<gui::LockedInfoWindow>(app);
+        });
+        windowsFactory.attach(desktop_reboot, [](Application *app, const std::string newname) {
+            return std::make_unique<gui::RebootWindow>(app);
+        });
     }
 
     void ApplicationDesktop::destroyUserInterface()