+++ module-apps/windows/AppWindow.hpp
@@ -45,9 +45,9 @@ namespace gui
         app::Application *application = nullptr;
 
       public:
+        AppWindow() = delete;
         AppWindow(app::Application *app, std::string name);
-        AppWindow(AppWindow *win);
-        virtual ~AppWindow();
+        ~AppWindow() override;
 
         app::Application *getApplication()
         {