+++ module-db/Interface/ThreadRecord.cpp
@@ -6,6 +6,7 @@
 #include <queries/messages/threads/QueryThreadGetByNumber.hpp>
 #include <queries/messages/threads/QueryThreadGetByContactID.hpp>
 #include <queries/messages/threads/QueryThreadRemove.hpp>
+#include <queries/messages/threads/QueryThreadsGet.hpp>
 
 #include <cassert>
 #include <log/log.hpp>
@@ -83,7 +84,7 @@ std::unique_ptr<std::vector<ThreadRecord>> ThreadRecordInterface::GetLimitOffset
 {
     auto records = std::make_unique<std::vector<ThreadRecord>>();
 
-    ThreadsTableFields threadsField;
+    ThreadsTableFields threadsField = ThreadsTableFields();
     switch (field) {
     case ThreadRecordField::ContactID: {
         threadsField = ThreadsTableFields::ContactID;