+++ module-utils/i18/i18.hpp
@@ -165,7 +165,10 @@ namespace utils
 
     // Global instance of i18 class
     extern i18 localize;
-    auto translateI18 = [](const std::string &text) { return utils::localize.get(text); };
+    inline auto translateI18(const std::string &text)
+    {
+        return utils::localize.get(text);
+    };
 } // namespace utils
 
 #endif // UNTITLED_I18_HPP